backend/requirements.txt
authsignal
flask
flask-restx
python-decouple
flask-cors